requirements.txt
bcrypt==4.2.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==2.0.1
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==2.0.1
WTForms==3.1.2
